chore: move tree repo to plugin
