+++ packages/core/database/src/index.ts
@@ -55,4 +55,3 @@ export * from './helpers';
 export { default as sqlParser, SQLParserTypes } from './sql-parser';
 export * from './interfaces';
 export { default as fieldTypeMap } from './view/field-type-map';
-export * from './repositories/tree-repository/adjacency-list-repository';
